🚚 Move wallet initialization screens files
